+++ plugins/tt-agentboard/test/services/workflow-runner.test.ts
@@ -1,29 +1,9 @@
-import { describe, it, expect, vi, beforeEach } from "vitest";
+import { describe, it, expect } from "vitest";
 import {
   checkPassCondition,
   renderTemplate,
   shellEscape,
 } from "../../server/domains/execution/workflow-helpers";
-
-import {
-  WorkflowRunner,
-  resolveStepComplete,
-} from "../../server/domains/execution/workflow-runner";
-import {
-  createMockDb,
-  createMockEventBus,
-  createMockLogger,
-  createMockTmuxManager,
-  createMockSlotAllocator,
-  createMockWorkflowLoader,
-  createMockContextBundler,
-  createMockStreamTailer,
-  createMockExecSync,
-  createMockCardService,
-  setupSelectReturning,
-  setupUpdate,
-} from "../helpers/mock-deps";
-import type { MockDb } from "../helpers/mock-deps";
 
 describe("workflow-helpers", () => {
   describe("checkPassCondition()", () => {
@@ -111,159 +91,6 @@ describe("workflow-helpers", () => {
 
     it("does not escape double quotes", () => {
       expect(shellEscape('say "hi"')).toBe("'say \"hi\"'");
-    });
-  });
-});
-
-describe("resolveStepComplete()", () => {
-  it("returns false when no pending callback exists", () => {
-    expect(resolveStepComplete(999)).toBe(false);
-  });
-});
-
-describe("WorkflowRunner", () => {
-  let runner: WorkflowRunner;
-  let mockDb: MockDb;
-  let mockEventBus: ReturnType<typeof createMockEventBus>;
-  let mockTmuxManager: ReturnType<typeof createMockTmuxManager>;
-  let mockSlotAllocator: ReturnType<typeof createMockSlotAllocator>;
-  let mockWorkflowLoader: ReturnType<typeof createMockWorkflowLoader>;
-  let mockCardService: ReturnType<typeof createMockCardService>;
-
-  beforeEach(() => {
-    mockDb = createMockDb();
-    mockEventBus = createMockEventBus();
-    mockTmuxManager = createMockTmuxManager();
-    mockSlotAllocator = createMockSlotAllocator();
-    mockWorkflowLoader = createMockWorkflowLoader();
-    mockCardService = createMockCardService();
-
-    runner = new WorkflowRunner(4200, {
-      db: mockDb as never,
-      eventBus: mockEventBus,
-      logger: createMockLogger(),
-      tmuxManager: mockTmuxManager,
-      slotAllocator: mockSlotAllocator as never,
-      workflowLoader: mockWorkflowLoader as never,
-      contextBundler: createMockContextBundler(),
-      writeHooks: vi.fn(),
-      cardService: mockCardService as never,
-      streamTailer: createMockStreamTailer(),
-      execSync: createMockExecSync() as never,
-      existsSync: vi.fn().mockReturnValue(true) as never,
-      readFileSync: vi.fn().mockReturnValue("PASS\ndetails") as never,
-    });
-    vi.clearAllMocks();
-  });
-
-  describe("initContext (via run)", () => {
-    it("returns early when card not found", async () => {
-      setupSelectReturning(mockDb, []);
-
-      await runner.run(999);
-
-      // Should not emit any workflow events since card not found
-      expect(mockEventBus.emit).not.toHaveBeenCalledWith("workflow:completed", expect.anything());
-    });
-
-    it("marks failed when card has no repoId", async () => {
-      const card = { id: 1, repoId: null, workflowId: "plan", title: "Test" };
-      setupSelectReturning(mockDb, [card]);
-      setupUpdate(mockDb);
-
-      await runner.run(1);
-
-      expect(mockCardService.updateStatus).toHaveBeenCalledWith(1, "failed");
-    });
-
-    it("marks failed when card has no workflowId", async () => {
-      const card = { id: 1, repoId: 1, workflowId: null, title: "Test" };
-      setupSelectReturning(mockDb, [card]);
-      setupUpdate(mockDb);
-
-      await runner.run(1);
-
-      expect(mockCardService.updateStatus).toHaveBeenCalledWith(1, "failed");
-    });
-
-    it("marks failed when workflow not found", async () => {
-      const card = { id: 1, repoId: 1, workflowId: "nonexistent", title: "Test" };
-      const repo = { id: 1, name: "test-repo", org: "org", defaultBranch: "main" };
-
-      let selectCount = 0;
-      mockDb.select = vi.fn().mockImplementation(() => {
-        selectCount++;
-        const chain: Record<string, unknown> = {};
-        chain.from = vi.fn().mockReturnValue(chain);
-        if (selectCount === 1) {
-          chain.where = vi.fn().mockResolvedValue([card]);
-        } else {
-          chain.where = vi.fn().mockResolvedValue([repo]);
-        }
-        return chain;
-      });
-
-      mockWorkflowLoader.get.mockReturnValue(undefined as never);
-      setupUpdate(mockDb);
-
-      await runner.run(1);
-
-      expect(mockCardService.updateStatus).toHaveBeenCalledWith(1, "failed");
-    });
-
-    it("marks failed when tmux not available", async () => {
-      const card = { id: 1, repoId: 1, workflowId: "plan", title: "Test" };
-      const repo = { id: 1, name: "test-repo", org: "org", defaultBranch: "main" };
-      const workflow = { name: "plan", steps: [], branch_template: "ab/card-{card_id}" };
-
-      let selectCount = 0;
-      mockDb.select = vi.fn().mockImplementation(() => {
-        selectCount++;
-        const chain: Record<string, unknown> = {};
-        chain.from = vi.fn().mockReturnValue(chain);
-        if (selectCount === 1) {
-          chain.where = vi.fn().mockResolvedValue([card]);
-        } else {
-          chain.where = vi.fn().mockResolvedValue([repo]);
-        }
-        return chain;
-      });
-
-      mockWorkflowLoader.get.mockReturnValue(workflow as never);
-      mockTmuxManager.isAvailable.mockReturnValue(false);
-      setupUpdate(mockDb);
-
-      await runner.run(1);
-
-      expect(mockCardService.updateStatus).toHaveBeenCalledWith(1, "failed");
-    });
-
-    it("marks queued when no slots available", async () => {
-      const card = { id: 1, repoId: 1, workflowId: "plan", title: "Test" };
-      const repo = { id: 1, name: "test-repo", org: "org", defaultBranch: "main" };
-      const workflow = { name: "plan", steps: [], branch_template: "ab/card-{card_id}" };
-
-      let selectCount = 0;
-      mockDb.select = vi.fn().mockImplementation(() => {
-        selectCount++;
-        const chain: Record<string, unknown> = {};
-        chain.from = vi.fn().mockReturnValue(chain);
-        if (selectCount === 1) {
-          chain.where = vi.fn().mockResolvedValue([card]);
-        } else {
-          chain.where = vi.fn().mockResolvedValue([repo]);
-        }
-        return chain;
-      });
-
-      mockWorkflowLoader.get.mockReturnValue(workflow as never);
-      mockTmuxManager.isAvailable.mockReturnValue(true);
-      mockSlotAllocator.claimSlot.mockResolvedValue(null);
-      setupUpdate(mockDb);
-
-      await runner.run(1);
-
-      expect(mockCardService.updateStatus).toHaveBeenCalledWith(1, "queued");
     });
   });
 });
